+++ fittrackee/tests/users/test_auth_api.py
@@ -558,7 +558,7 @@ class TestUserProfile(ApiTestCaseMixin):
         assert data['data']['imperial_units'] is False
         assert data['data']['nb_sports'] == 2
         assert data['data']['nb_workouts'] == 2
-        assert len(data['data']['records']) == 6
+        assert len(data['data']['records']) == 8
         assert data['data']['sports_list'] == [1, 2]
         assert data['data']['total_distance'] == 22
         assert data['data']['total_duration'] == '2:40:00'
@@ -1112,7 +1112,6 @@ class TestUserPicture(ApiTestCaseMixin):
         )
 
         data = json.loads(response.data.decode())
-        print('data', data)
         assert response.status_code == 413
         assert 'fail' in data['status']
         assert (
@@ -1144,7 +1143,6 @@ class TestUserPicture(ApiTestCaseMixin):
         )
 
         data = json.loads(response.data.decode())
-        print('data', data)
         assert response.status_code == 413
         assert 'fail' in data['status']
         assert (