+++ fittrackee_client/src/components/Administration/AdminMenu.vue
@@ -65,7 +65,7 @@
 </script>
 
 <style lang="scss" scoped>
-  @import '~@/scss/base.scss';
+  @import '~@/scss/vars.scss';
 
   #admin-menu {
     display: flex;