+++ fittrackee_client/src/components/Common/Error.vue
@@ -26,8 +26,6 @@
 </script>
 
 <style scoped lang="scss">
-  @import '~@/scss/base';
-
   #error {
     display: flex;
     align-items: center;