+++ fittrackee_client/src/components/Common/Modal.vue
@@ -58,7 +58,7 @@
 </script>
 
 <style lang="scss" scoped>
-  @import '~@/scss/base';
+  @import '~@/scss/vars.scss';
   #modal {
     position: fixed;
     top: 0;