+++ fittrackee_client/src/components/Dashboard/UserCalendar/CalendarHeader.vue
@@ -36,7 +36,7 @@
 </script>
 
 <style lang="scss">
-  @import '~@/scss/base';
+  @import '~@/scss/vars.scss';
   .calendar-header {
     display: flex;
     flex-direction: row;