+++ fittrackee_client/src/components/Dashboard/UserMonthStats.vue
@@ -40,7 +40,7 @@
 </script>
 
 <style lang="scss" scoped>
-  @import '~@/scss/base';
+  @import '~@/scss/vars.scss';
 
   .user-month-stats {
     ::v-deep(.card-content) {