+++ fittrackee_client/src/views/AdminView.vue
@@ -39,7 +39,7 @@
 </script>
 
 <style lang="scss" scoped>
-  @import '~@/scss/base.scss';
+  @import '~@/scss/vars.scss';
 
   #admin {
     .admin-card {