+++ fittrackee_client/src/views/user/UserView.vue
@@ -36,7 +36,7 @@
 </script>
 
 <style lang="scss" scoped>
-  @import '~@/scss/base.scss';
+  @import '~@/scss/vars.scss';
 
   #user {
     margin: auto;